* Makefile.in (t_walk_rtree, t_kerb): Do not link against kdb libraries 	for these test executables.
	* srv_rcache.c (krb5_get_server_rcache): Cast argument to
	isgraph() to int.

	* init_ctx.c: Cast arguments to isspace() to int. If unix is defined,
	include ../krb5_libinit.h. There has to be a better was for windows.

	* conv_princ.c (krb5_425_conv_principal): Cast argument to isupper().
	to int.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13340 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/krb/conv_princ.c
@@ -318,7 +318,7 @@ krb5_425_conv_principal(context, name, instance, realm, princ)
 		      return retval;
 		  if (domain) {
 		      for (cp = domain; *cp; cp++)
-			  if (isupper(*cp))
+			  if (isupper((int) (*cp)))
 			      *cp = tolower(*cp);
 		      strncat(buf, ".", sizeof(buf) - 1 - strlen(buf));
 		      strncat(buf, domain, sizeof(buf) - 1 - strlen(buf));
